feat(routing): add @ngrx/router-store
For now we will use it to dispatch route changes via store.dispatch(),
while we (usually) read the state using Angular’s ActivatedRoute.

Not yet entirely sure what are the benefits of this extra layer
(apart of central store from where the route state can be easily read),
but let’s try it for now.